Add files via upload
Signal waveform but for 1kHz (counter only goes up to 49_999). This was a lazy way to reduce the amount of points to simulate and still check is the circuit works. If you can't see some of the signals, click on "dut" (the name of counter instance from the simulation file) find a needed signal in "Objects" and drag'n'drop in to the field with other waveform names